chore: Update AdminController and AuthController to load additional data
This commit updates the AdminController and AuthController files in the Back-End directory. In the AdminController, the "dashboard" method now loads additional data by calling the "load" method on the "transaksi_terakhir" variable. In the AuthController, the "login" method now includes the user's role in the JWT payload. These changes were made to improve the functionality and user experience of the application.